+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/storage/annotation/BanyanDB.java
@@ -134,6 +134,13 @@ public @interface BanyanDB {
 
     /**
      * MeasureField defines a column as a measure's field.
+     *
+     * Annotated: the column is a measure field.
+     * Unannotated: the column is a measure tag.
+     *   storageOnly=true: the column is a measure tag which is not indexed.
+     *   storageOnly=false: the column is a measure tag which is indexed.
+     *     indexOnly=true: the column is a measure tag which is indexed, but not stored.
+     *     indexOnly=false: the column is a measure tag which is indexed and stored.
      * @since 9.4.0
      */
     @Target({ElementType.FIELD})