+++ docs/en/changes/changes.md
@@ -73,6 +73,8 @@
 * Remove `id` tag from measures.
 * Add `Banyandb.MeasureField` to mark a column as a BanyanDB Measure field.
 * Add `BanyanDB.StoreIDTag` to store a process's id for searching.
+* [**Breaking Change**] The supported version of ShardingSphere-Proxy is upgraded from 5.1.2 to 5.3.1. Due to the changes of ShardingSphere's API, versions before 5.3.1 are not compatible.
+* Add the eBPF network profiling E2E Test in the per storage.
 
 #### UI